requirements.txt
openai==1.55.0
python-dotenv==1.0.1
GitPython==3.1.43
alembic==1.14.0
SQLAlchemy==2.0.36
pydantic-settings==2.6.1
pydantic==2.10.1
sqlmodel==0.0.22
psycopg2==2.9.10
pika==1.3.2
fastapi==0.115.5
fastapi-route-logger-middleware==0.2.3
sqlalchemy-pagination==0.0.2
uvicorn==0.32.1
